Split XCTarget loader from reference
